requirements.txt
asgiref==3.2.7
cx-Oracle==7.3.0
cycler==0.10.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
kiwisolver==1.2.0
matplotlib==3.2.1
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.4
pandas==1.0.3
pycryptodome==3.9.7
pydot==1.4.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
sqlalchemy-diff==0.1.3
sqlalchemy-schemadisplay==1.3
SQLAlchemy-Utils==0.36.6
sqlparse==0.3.1
teradatasql==16.20.0.62
teradatasqlalchemy==16.20.0.8
